medgemma/medgemma-app/requirements.txt
# Core ML/AI
torch>=2.0.0
transformers>=4.50.0
accelerate>=0.26.0
Pillow>=10.0.0
huggingface_hub>=0.20.0
sentencepiece>=0.1.99
bitsandbytes>=0.41.0

# Backend API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
httpx>=0.27.0
aiofiles>=23.2.1

# Frontend
streamlit>=1.32.0
streamlit-option-menu>=0.3.12
streamlit-extras>=0.4.0

# Data & Utilities
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.18.0
scipy>=1.11.0
requests>=2.31.0
python-dotenv>=1.0.0

# Medical / DICOM support
pydicom>=2.4.0
SimpleITK>=2.3.0

# PDF/Report generation
reportlab>=4.0.0
fpdf2>=2.7.0

# Audio (for future MedASR integration)
soundfile>=0.12.1

# Logging & monitoring
loguru>=0.7.0
